changes: added service file and pagination
