+++ validation/categoryValidation.js
@@ -29,6 +29,9 @@ module.exports = {
     condition: Joi.object().optional().messages({
       'object.base': 'Condition should be type of an object',
     }),
+    page: Joi.number().optional().messages({
+      'number.base': 'page should be type of a number',
+    }),
     pageSize: Joi.number().optional().messages({
       'number.base': 'pageSize should be type of a number',
     }),
